# improve script events v2