--- a/plugin/src/main/java/net/aufdemrand/denizen/events/BukkitScriptEvent.java
+++ b/plugin/src/main/java/net/aufdemrand/denizen/events/BukkitScriptEvent.java
@@ -6,6 +6,7 @@ import net.aufdemrand.denizencore.events.ScriptEvent;
 import net.aufdemrand.denizencore.scripts.containers.ScriptContainer;
 import net.aufdemrand.denizencore.utilities.CoreUtilities;
 import org.bukkit.Location;
+import org.bukkit.Material;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Hanging;
 import org.bukkit.entity.Projectile;
@@ -108,16 +109,16 @@ public abstract class BukkitScriptEvent extends ScriptEvent {
         if (comparedto.equals("item")) {
             return true;
         }
-        if (comparedto.equals("potion") && CoreUtilities.toLowerCase(item.getItemStack().getType().name()).contains("potion")) {
+        else if (comparedto.equals("potion") && CoreUtilities.toLowerCase(item.getItemStack().getType().name()).contains("potion")) {
             return true;
+        }
+        Material material = Material.getMaterial(comparedto);
+        if (material == null || item.getMaterial().getMaterial() != material) {
+        	return false;
         }
         item = new dItem(item.getItemStack().clone());
         item.setAmount(1);
-        dMaterial material = dMaterial.valueOf(comparedto);
-        if (material == null || item.getMaterial().getMaterial() != material.getMaterial()) {
-        	return false;
-        }
-        else if (CoreUtilities.toLowerCase(item.identifyNoIdentifier()).equals(comparedto)) {
+        if (CoreUtilities.toLowerCase(item.identifyNoIdentifier()).equals(comparedto)) {
             return true;
         }
         else if (CoreUtilities.toLowerCase(item.identifyMaterialNoIdentifier()).equals(comparedto)) {
@@ -138,12 +139,12 @@ public abstract class BukkitScriptEvent extends ScriptEvent {
             return false;
         }
         comparedto = CoreUtilities.toLowerCase(comparedto);
-        dMaterial material = dMaterial.valueOf(comparedto);
-        if (material == null || mat.getMaterial() != material.getMaterial()) {
-        	return false;
-        }
-        else if (comparedto.equals("block") || comparedto.equals("material")) {
+        if (comparedto.equals("block") || comparedto.equals("material")) {
             return true;
+        }
+        Material material = Material.getMaterial(comparedto);
+        if (material == null || mat.getMaterial() != material) {
+        	return false;
         }
         else if (CoreUtilities.toLowerCase(mat.realName()).equals(comparedto)) {
             return true;