Aight. _on player clicks block_ didn't work in my last PR, as the new checks were before _.equals("block")_, 
    so I changed the order here to fix that.

Working with Material.getMaterial() directly seems even faster, as it uses final/static references to lookup (dMaterial iterates with String checks instead).

**placing of stone:**
_without: 0.29%	2.35%	0.02 s	1.17 ms
last PR: 0.05%	0.48%	0.00 s	0.24 ms
this PR: 0.02%	   0.14%	  0.00 s	     0.07 ms_

You may look again, for possible conflicts ?