@mcmonkey4eva 2018-06-23
Does not work with dItems or dMaterials...